pyproject.toml
[tool.poetry]
name = "development-task"
version = "0.1.0"
description = ""
authors = ["prasanta <prasantakakati1994@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
certifi = "2017.4.17"
chardet = "3.0.4"
defusedxml = "0.5.0"
dj-database-url = "0.4.2"
Django = "1.10"
django-bootstrap3 = "7.0.1"
django-braces = "1.10.0"
django-filer = "0.8.7"
django-js-asset = "1.1.0"
django-mptt = "0.9.0"
django-oauth-toolkit = "0.11.0"
django-polymorphic = "0.5.5"
django-rest-framework-social-oauth2 = "1.0.5"
djangorestframework = "3.4.3"
easy-thumbnails = "2.5"
gunicorn = "19.6.0"
idna = "2.5"
oauthlib = "1.1.2"
olefile = "0.44"
six = "1.10.0"
social-auth-app-django = "1.1.0"
social-auth-core = "1.7.0"
urllib3 = "1.21.1"
Pillow = "^7.0.0"
psycopg2 = "2.7.3.2"
PyJWT = "1.5.0"
python-social-auth = "0.3.6"
python3-openid = "3.1.0"
pytz = "2016.6.1"
requests = "2.17.3"
requests-oauthlib = "0.8.0"
shortuuid = "0.5.0"
stripe = "1.37.0"
whitenoise = "3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
